Make CS consistently active high.
Existing code passed CS mostly as active low, except for the master CSR,
but the signal naming was inconsistent on whether it had the _n suffix or not.

This patch makes CS active high everywhere except for the output to the pad
itself and updates the signal names to match.