phy/generic: make minimum CS inactive time configurable
If LiteSPIPHY is used in a design where the system clock is relatively
slow and the minimum CS high time is known for a specific flash chip, it
can be desirable to pass in a cs_delay lower than 10.

For example, with a system clock of 75MHz and using GD25LQ128D which
specifies 20ns minimum CS high time, the delay can be satisfied in only
2 clock cycles instead of 10. This gives a slight throughput improvement
for non-sequential access to flash.