mmap: Add write functionality
With this change, it is possible to memory map e.g. PSRAM. When setting
with_write to "csr", it can also be used to support faster flash writing
via etherbone. For flash, only False and "csr" should be used, as the
some other logic or application needs to make sure that the flash is
erased ahead of time. True is only meant for use with RAM chips.

Signed-off-by: Matthias Breithaupt <m.breithaupt@vogl-electronic.com>