core: be able to set clk dividers independent per core
be able to set clk dividers independent per core.
When using the master sometimes we need a
slower speed, in order to not change it for mmap
and make that slower too, they now can be set per
core module. When the clk divider is not set in the
core module (by it being 0), the phy will use the
 clk divider it's own csr.

Signed-off-by: Fin Maaß <f.maass@vogl-electronic.com>